packages/workers-py/requirements.txt
Pillow>=10.0.0
scikit-learn>=1.3.0
imagehash>=4.3.0
rembg>=2.0.0
playwright>=1.40.0
bullmq>=2.0.0
redis>=5.0.0
boto3>=1.34.0
psycopg2-binary>=2.9.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
